Split src/runtime in to src/runtime and src/loader files. This will make the architecture clearer, improve make rules.
