+++ .gitignore
@@ -24,7 +24,7 @@ src/outputgeneration/SourceMapIntegration.js
 src/syntax/ParseTreeVisitor.js
 src/syntax/trees/ParseTreeType.js
 src/syntax/trees/ParseTrees.js
-src/runtime/version.js
+src/loader/version.js
 test/unit/node/out/compile-amd
 test/unit/node/out/compile-cjs
 test/unit/node/out/compile-cjs-maps
