+++ src/loader/InlineLoaderCompiler.js
