+++ src/loader/ModuleStore.js
