+++ src/loader/ModuleStoreImpl.js
