+++ src/loader/NodeTraceurLoader.js
