+++ src/loader/TraceurLoader.js
